ISSUE #3738 - rename duedate label files and components for consistency
